requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
click>=8.1.0
colorama>=0.4.6
rich>=13.7.0
python-dotenv>=1.0.0
regex>=2023.0.0
